Fix compiler warnings due to -Wextra
Add -Wextra to Makefile

Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/186726
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>

+++ benchmarks/bm_lock.c
@@ -689,7 +689,7 @@ usage :
 	    break;
 	}
     }
-    if (LOCKTYPE == -1)
+    if ((int)LOCKTYPE == -1)
     {
 	goto usage;
     }
